+++ examples/room-manager/src/roomService.ts
